Refactor args processing and generator entrypoint
Remove args, minimally process in main. Add generator struct with Run
method as central entrypoint. Provide config with metadata to generator
via constructor function. Add metadata with a plain constructor and a
"decorating" validator function which is used in main. The plain
constructor is used in tests. Refactored the e2e test for creating docs
to use the generator.