Fix top-k-frequent spread operator error
The spread operator in the final for loop lead to wrong returns. With
k=1 and an input of [1, 2] the spread operator lead to a result of
[1, 2] instead of just [1]. Instead of the spread operator a second loop
was needed to iterate over the nested array and extract only k elements
and not all. A test was added to cover the case.