Fix top-k-frequent non deterministic test case
When a map is used it is not guaranteed that the keys of the map are
always in the same order when they are iterated. In the new test case
where k=1 and the input is [1,2,3] there are three correct results
namely [1], [2], [3]. The TopKFrequent function returns one of those
correct results but not always the same. The failing test case was
always checking for one of those three numbers and sometimes failed. I
extracted the test case into a new unit test and check that the result
is one of the correct numbers.